v0.2.0-beta
# 0.2.0-beta - 2024-02-28

<small>[Compare to previous release][comp:0.2.0-beta]</small>

### Breaking Changes

-   Moved the `Connector` and `Matches` enumerators from `TopMarksDevelopment.ExpressionBuilder.Api` to `TopMarksDevelopment.ExpressionBuilder`, for easier access

### Fixes

- Removed duplicated methods that we in both the `TopMarksDevelopment.ExpressionBuilder.Api` and `TopMarksDevelopment.ExpressionBuilder` namespaces. Preventing you from using an extension method if you use both namespaces

### Changes

- Added GitHub `FUNDING.yml` to hopefully get some support/funding 😜
- (Non-user facing) Updated the GitHub workflows to use the current versions of `checkout` and `setup-dotnet`

[comp:0.2.0-beta]: https://github.com/TopMarksDevelopment/Expression-Builder/compare/v0.2.0-beta...v0.1.0-beta
