+++ .github/FUNDING.yml
@@ -1,0 +1,5 @@
+# There is absolutely no pressure here. If you want to support me then that's great!
+# It will help cover hosting costs and the hours I have invested - plus it will put a little spring in my step 😁
+
+github: [TopMarksDevelopment]
+custom: ['https://www.paypal.com/donate?hosted_button_id=5UFB62RB4AWNL']
