+++ .github/workflows/build-and-test.yml
@@ -19,7 +19,7 @@ jobs:
                 os: [ubuntu-latest, windows-latest, macOS-latest]
 
         steps:
-            - uses: actions/checkout@v3
+            - uses: actions/checkout@v4
             - name: Setup .NET Core
               uses: actions/setup-dotnet@v3
               with:
