+++ .github/workflows/push.yml
@@ -13,10 +13,10 @@ jobs:
         runs-on: windows-latest
 
         steps:
-            - uses: actions/checkout@v3
+            - uses: actions/checkout@v4
 
             - name: Setup .NET
-              uses: actions/setup-dotnet@v3
+              uses: actions/setup-dotnet@v4
               with:
                   dotnet-version: 8.0.x
