--- Packages/Core/src/Extensions/IFilterableAddExtensions.cs
@@ -1,169 +1,0 @@
-namespace TopMarksDevelopment.ExpressionBuilder;
-
-using System;
-using System.Linq.Expressions;
-using TopMarksDevelopment.ExpressionBuilder.Api;
-
-public static partial class IFilterableEqualExtensions
-{
-    #region Expression<Func<,>> propertyExpression
-
-    public static IFilterable Add<TClass, TPropertyType>(
-        this IFilterable filter,
-        Expression<Func<TClass, TPropertyType?>> propertyExpression,
-        IOperation operation,
-        TPropertyType?[] value,
-        IFilterStatementOptions? options,
-        Connector connector
-    )
-        where TClass : class
-    {
-        var f = filter.Add(propertyExpression, operation, value, options);
-
-        return connector == Connector.Or ? f.Or() : f.And();
-    }
-
-    public static IFilterable Add<TClass, TPropertyType>(
-        this IFilterable filter,
-        Expression<Func<TClass, TPropertyType?>> propertyExpression,
-        IOperation operation,
-        TPropertyType? value,
-        IFilterStatementOptions? options,
-        Connector connector
-    )
-        where TClass : class =>
-        filter.Add(propertyExpression, operation, [value], options, connector);
-
-    public static IFilterableConnection Add<TClass, TPropertyType>(
-        this IFilterable filter,
-        Expression<Func<TClass, TPropertyType?>> propertyExpression,
-        IOperation operation,
-        TPropertyType?[] value,
-        IFilterStatementOptions? options
-    )
-        where TClass : class =>
-        filter.Add(propertyExpression, operation, value, options);
-
-    public static IFilterableConnection Add<TClass, TPropertyType>(
-        this IFilterable filter,
-        Expression<Func<TClass, TPropertyType?>> propertyExpression,
-        IOperation operation,
-        TPropertyType? value,
-        IFilterStatementOptions? options
-    )
-        where TClass : class =>
-        filter.Add(propertyExpression, operation, [value], options);
-
-    public static IFilterableConnection Add<TClass, TPropertyType>(
-        this IFilterable filter,
-        Expression<Func<TClass, TPropertyType?>> propertyExpression,
-        IOperation operation,
-        TPropertyType?[] value
-    )
-        where TClass : class =>
-        filter.Add(propertyExpression, operation, value, null);
-
-    public static IFilterableConnection Add<TClass, TPropertyType>(
-        this IFilterable filter,
-        Expression<Func<TClass, TPropertyType?>> propertyExpression,
-        IOperation operation,
-        TPropertyType? value
-    )
-        where TClass : class =>
-        filter.Add(propertyExpression, operation, [value]);
-
-    public static IFilterable Add<TClass, TPropertyType>(
-        this IFilterable filter,
-        Expression<Func<TClass, TPropertyType?>> propertyExpression,
-        IOperation operation,
-        TPropertyType?[] value,
-        Connector connector
-    )
-        where TClass : class =>
-        filter.Add(propertyExpression, operation, value, null, connector);
-
-    public static IFilterable Add<TClass, TPropertyType>(
-        this IFilterable filter,
-        Expression<Func<TClass, TPropertyType?>> propertyExpression,
-        IOperation operation,
-        TPropertyType? value,
-        Connector connector
-    )
-        where TClass : class =>
-        filter.Add(propertyExpression, operation, [value], null, connector);
-
-    #endregion Expression<Func<,>> propertyExpression
-
-    #region string propertyExpression
-
-    public static IFilterable Add<TPropertyType>(
-        this IFilterable filter,
-        string propertyExpression,
-        IOperation operation,
-        TPropertyType?[] value,
-        IFilterStatementOptions? options,
-        Connector connector
-    )
-    {
-        var f = filter.Add(propertyExpression, operation, value, options);
-
-        return connector == Connector.Or ? f.Or() : f.And();
-    }
-
-    public static IFilterable Add<TPropertyType>(
-        this IFilterable filter,
-        string propertyExpression,
-        IOperation operation,
-        TPropertyType? value,
-        IFilterStatementOptions? options,
-        Connector connector
-    ) => filter.Add(propertyExpression, operation, [value], options, connector);
-
-    public static IFilterableConnection Add<TPropertyType>(
-        this IFilterable filter,
-        string propertyExpression,
-        IOperation operation,
-        TPropertyType?[] value,
-        IFilterStatementOptions? options
-    ) => filter.Add(propertyExpression, operation, value, options);
-
-    public static IFilterableConnection Add<TPropertyType>(
-        this IFilterable filter,
-        string propertyExpression,
-        IOperation operation,
-        TPropertyType? value,
-        IFilterStatementOptions? options
-    ) => filter.Add(propertyExpression, operation, [value], options);
-
-    public static IFilterableConnection Add<TPropertyType>(
-        this IFilterable filter,
-        string propertyExpression,
-        IOperation operation,
-        TPropertyType?[] value
-    ) => filter.Add(propertyExpression, operation, value, null);
-
-    public static IFilterableConnection Add<TPropertyType>(
-        this IFilterable filter,
-        string propertyExpression,
-        IOperation operation,
-        TPropertyType? value
-    ) => filter.Add(propertyExpression, operation, [value], null);
-
-    public static IFilterable Add<TPropertyType>(
-        this IFilterable filter,
-        string propertyExpression,
-        IOperation operation,
-        TPropertyType?[] value,
-        Connector connector
-    ) => filter.Add(propertyExpression, operation, value, null, connector);
-
-    public static IFilterable Add<TPropertyType>(
-        this IFilterable filter,
-        string propertyExpression,
-        IOperation operation,
-        TPropertyType? value,
-        Connector connector
-    ) => filter.Add(propertyExpression, operation, [value], null, connector);
-
-    #endregion string propertyExpression
-}
